+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Fix zstd decompression to read across frames.
   ([#6921](https://github.com/mitmproxy/mitmproxy/pull/6921), @zendai)
+* Add `HttpConnectedHook` and `HttpConnectErrorHook`
 
 ## 12 June 2024: mitmproxy 10.3.1